move unit tests to test/easyblocks
