+++ test/easyblocks/init_easyblocks.py
